requirements.txt
attrs==21.4.0
click==8.1.3
dnspython==2.2.1
flasgger==0.9.5
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
Flask-PyMongo==2.3.0
gunicorn==20.1.0
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
MarkupSafe==2.1.1
mistune==2.0.2
PyJWT==2.4.0
pymongo==4.1.1
pyrsistent==0.18.1
python-dotenv==0.20.0
PyYAML==6.0
six==1.16.0
Werkzeug==2.1.2
zipp==3.8.0
